requirements.txt
requests==2.22.0
eth-abi==2.1.0
pysha3==1.0.2
ecdsa==0.15
secp256k1==0.13.2
protobuf==3.11.3
